MOve the chat bot to a single component and fix the formatting
